[refactor](SqlExecWorkflow): SQL audit result list operation permission[skip ci]
